Refactor canister 'ic_eth' -> 'eth_util'
