+++ canisters/eth_util/src/lib.rs
@@ -1,20 +1,30 @@
 use std::{rc::Rc, str::FromStr};
 
 use candid::candid_method;
-use eth_rpc::call_contract;
-use ethers_core::{
+use ic_eth::core::{
     abi::{Contract, Token},
     types::{Address, RecoveryMessage, Signature},
-};
-use util::to_hex;
 
-mod eth_rpc;
-mod util;
+use ic_eth;;::call_contract;
+use ic_eth::util::to_hex;
+
+const HTTP_CYCLES: u128 = 100_000_000;
+const MAX_RESPONSE_BYTES: u64 = 1000;
 
 // Load relevant ABIs (Ethereum equivalent of Candid interfaces)
 thread_local! {
     static ERC_721: Rc<Contract> = Rc::new(include_abi!("../abi/erc721.json"));
     static ERC_1155: Rc<Contract> = Rc::new(include_abi!("../abi/erc1155.json"));
+}
+
+/// Choose the relevant JSON-RPC endpoint for the given network.
+fn get_rpc_endpoint(network: &str) -> &'static str {
+    match network {
+        "mainnet" | "ethereum" => "https://cloudflare-eth.com/v1/mainnet",
+        "goerli" => "https://ethereum-goerli.publicnode.com",
+        "sepolia" => "https://rpc.sepolia.org",
+        _ => panic!("Unsupported network: {}", network),
+    }
 }
 
 /// Verify an ECDSA signature (message signed by an Ethereum wallet).
@@ -39,11 +49,13 @@ pub async fn erc721_owner_of(network: String, contract_address: String, token_id
 
     let abi = &ERC_721.with(Rc::clone);
     let result = call_contract(
-        &network,
+        get_rpc_endpoint(&network),
         contract_address,
         abi,
         "ownerOf",
         &[Token::Uint(token_id.into())],
+        HTTP_CYCLES,
+        Some(MAX_RESPONSE_BYTES),
     )
     .await;
     match result.get(0) {
@@ -68,14 +80,16 @@ pub async fn erc1155_balance_of(
 
     let abi = &ERC_1155.with(Rc::clone);
     let result = call_contract(
-        &network,
+        get_rpc_endpoint(&network),
         contract_address,
-        abi,
+        *abi,
         "balanceOf",
         &[
             Token::Address(owner_address.into()),
             Token::Uint(token_id.into()),
         ],
+        HTTP_CYCLES,
+        Some(MAX_RESPONSE_BYTES),
     )
     .await;
     match result.get(0) {
