--- canisters/ic_eth/src/util.rs
@@ -1,9 +1,0 @@
-use hex::FromHexError;
-
-pub fn to_hex(data: &[u8]) -> String {
-    format!("0x{}", hex::encode(data))
-}
-
-pub fn from_hex(data: &str) -> Result<Vec<u8>, FromHexError> {
-    hex::decode(&data[2..])
-}
